% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{workingpaper}
\alias{workingpaper}
\title{Output formats for working papers}
\usage{
workingpaper(...)
}
\arguments{
\item{\dots}{Arguments passed to \code{\link[bookdown]{pdf_document2}}.}
}
\value{
An R Markdown output format object.
}
\description{
Each function is a wrapper for \code{\link[bookdown]{pdf_document2}} to
produce customized documents. Copyright monash Copyright (C) 2017 Rob J Hyndman.
Modified by Priyanga Dilini Talagala for personal use.
}
\author{
Rob J Hyndman (Modified by Priyanga Dilini Talagala for personal use.)
}
